+++ src/routes/users/{id}/settings.test.js
@@ -1,6 +1,6 @@
 const test = require('ava');
 
-const { setup } = require('../../../test/helpers/setup');
+const { setup } = require('../../../../test/helpers/setup');
 
 test.before(async t => {
     const { server, getTeamWithUser, getUser } = await setup({
